+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/blocks/RepairedSpawner.java
@@ -3,6 +3,7 @@ package io.github.thebusybiscuit.slimefun4.implementation.items.blocks;
 import java.util.ArrayList;
 import java.util.Collection;
 
+import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.bukkit.Material;
@@ -12,7 +13,10 @@ import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.slimefun4.api.events.BlockPlacerPlaceEvent;
+import io.github.thebusybiscuit.slimefun4.api.items.ItemSetting;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
+import io.github.thebusybiscuit.slimefun4.core.handlers.BlockUseHandler;
+import io.github.thebusybiscuit.slimefun4.utils.tags.SlimefunTag;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
@@ -27,11 +31,31 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
  */
 public class RepairedSpawner extends AbstractMonsterSpawner {
 
+    private final ItemSetting<Boolean> allowSpawnEggs = new ItemSetting<>("allow-spawn-eggs", true);
+
     @ParametersAreNonnullByDefault
     public RepairedSpawner(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
 
-        addItemHandler(new BlockPlaceHandler(true) {
+        addItemSetting(allowSpawnEggs);
+
+        addItemHandler(onInteract());
+        addItemHandler(onPlace());
+    }
+
+    @Nonnull
+    private BlockUseHandler onInteract() {
+        return e -> {
+            if (!allowSpawnEggs.getValue() && SlimefunTag.SPAWN_EGGS.isTagged(e.getItem().getType())) {
+                // Disallow spawn eggs from being used on Reinforced Spawners if disabled
+                e.cancel();
+            }
+        };
+    }
+
+    @Nonnull
+    private BlockPlaceHandler onPlace() {
+        return new BlockPlaceHandler(true) {
 
             @Override
             public void onPlayerPlace(BlockPlaceEvent e) {
@@ -45,6 +69,10 @@ public class RepairedSpawner extends AbstractMonsterSpawner {
 
             @ParametersAreNonnullByDefault
             private void onPlace(ItemStack item, BlockEvent e) {
+                /**
+                 * This may no longer be needed at some point but for legacy items
+                 * we still need to set the spawned EntityType manually
+                 */
                 if (e.getBlock().getType() == Material.SPAWNER) {
                     getEntityType(item).ifPresent(entity -> {
                         CreatureSpawner spawner = (CreatureSpawner) e.getBlock().getState();
@@ -53,7 +81,7 @@ public class RepairedSpawner extends AbstractMonsterSpawner {
                     });
                 }
             }
-        });
+        };
     }
 
     @Override